+++ src/tables/mod.rs
@@ -20,6 +20,7 @@ pub mod gvar;
 pub mod head;
 pub mod hhea;
 pub mod hmtx;
+pub mod hvar;
 pub mod kern;
 pub mod loca;
 pub mod maxp;